requirements.txt
appdirs==1.4.4
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
cattrs==1.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
comtypes==1.1.11
docopt==0.6.2
Flask==2.1.1
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
jaraco.context==4.1.1
Jinja2==3.1.1
joblib==1.1.0
Js2Py==0.71
keyboard==0.13.5
lxml==4.8.0
MarkupSafe==2.1.1
more-itertools==8.12.0
MouseInfo==0.1.3
nltk==3.7
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pipwin==0.5.2
PyAudio
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyparsing==3.0.8
pyperclip==1.8.2
pypiwin32==223
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==0.1.28
pySmartDL==1.3.4
pyttsx3==2.90
pytweening==1.0.4
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.3
pywikihow==0.5.7
pywin32==303
regex==2022.3.15
requests==2.27.1
requests-cache==0.9.3
six==1.16.0
soupsieve==2.3.2
SpeechRecognition==3.8.1
speedtest-cli==2.1.3
torch==1.11.0
tqdm==4.64.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
url-normalize==1.4.3
urllib3==1.26.9
Werkzeug==2.1.1
wikipedia==1.4.0
wolframalpha==5.0.0
xmltodict==0.12.0
zipp==3.8.0
